requirements.txt
# Django and REST framework
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.0
django-filter==23.3
django-allauth==0.57.0
dj-rest-auth==5.0.1
drf-yasg==1.21.7

# Database
psycopg2-binary==2.9.9
dj-database-url==2.1.0

# Environment variables
python-dotenv==1.0.0

# Authentication
djangorestframework-simplejwt==5.3.0

# Production
gunicorn==21.2.0
whitenoise==6.5.0

# Utilities
Pillow==10.1.0
requests==2.31.0
